# docs/msc2/api-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MSC 2 v1 API contract",
    "version": "v1",
    "description": "The versioned HTTP contract MSC 2's skeletal agent (Phase 2) and every later phase implement against. Assembled per docs/msc2/rolling-plan.md P2.8 from the P0.23/P0.30/P0.32 MSC 1 baseline (docs/msc2/api-baseline/openapi.json) plus this phase's own designs: the /v1/ namespace and ErrorDTO envelope (P2.4, versioning-and-errors.md), the operation routes and OperationDTO (P2.5, operation-model.md), the capabilities route (P2.6, capability-model.md), helpId placement (P2.2, helpid-contract.md), and the permission-category annotation on every route (P2.1, permission-vocabulary.csv). The console and operation-progress WebSocket channels (P2.7) are documented separately in docs/msc2/api-contract/websocket-v1.json, matching the baseline's own precedent (P0.24) of keeping WS channels out of the OpenAPI paths object. Everything sourced from P2.4/P2.5/P2.6/P2.2's own documents carries their Proposed/Confirmed status forward -- this file does not itself promote anything to Approved."
  },
  "paths": {
...
    "/v1/servers/import": {
      "post": {
        "summary": "Scan or import an existing server directory/zip/transfer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerImportRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "action=scan returns ServerImportScanResponseDTO; any other action (importExisting/importTransfer) returns ServerImportResultDTO \u2014 same status code, different body shape, dispatched on the request's own 'action' field.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ServerImportScanResponseDTO"
                    },
                    {
                      "$ref": "#/components/schemas/ServerImportResultDTO"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "missing_body / invalid_json / invalid_action / missing_source_path / invalid_path / display_name_required / backup_path_required (typed for every cause except missing_body/invalid_json (still Error))",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "invalid_body"
          },
          "404": {
            "description": "source_not_found (typed for every cause except missing_body/invalid_json (still Error))",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "not_found"
          },
          "409": {
            "description": "server_running (typed for every cause except missing_body/invalid_json (still Error))",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "conflict"
          },
          "500": {
            "description": "internal error (typed for every cause except missing_body/invalid_json (still Error))",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "internal_error"
          }
        },
        "x-notes": "Read from RemoteAPIServer+ComponentRoutes.swift handleImportServer and importStatus. req.action == \"scan\" routes to serverImportScanProvider; anything else routes to serverImportProvider.",
        "x-permission-category": "fleet"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServerImportRequestDTO": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "sourcePath": {
            "type": "string"
          },
          "importKind": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "serverType": {
            "type": "string"
          },
          "activeWorldName": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "maxPlayers": {
            "type": "integer"
          },
          "acceptEula": {
            "type": "boolean"
          },
          "enablePlayit": {
            "type": "boolean"
          },
          "transferMode": {
            "type": "string"
          },
          "backupPath": {
            "type": "string"
          },
          "javaPortOverrides": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "bedrockPortOverrides": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        },
        "required": [
          "action",
          "sourcePath"
        ]
      },
      "ServerImportWorldDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          },
          "dimensionsLabel": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "sizeBytes",
          "dimensionsLabel"
        ]
      },
      "ServerImportScanResponseDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "sourcePath": {
            "type": "string"
          },
          "isZip": {
            "type": "boolean"
          },
          "serverType": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "maxPlayers": {
            "type": "integer"
          },
          "eulaAccepted": {
            "type": "boolean"
          },
          "worlds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerImportWorldDTO"
            }
          },
          "defaultWorldName": {
            "type": "string"
          },
          "javaFlavor": {
            "type": "string"
          },
          "detectedMCVersion": {
            "type": "string"
          },
          "detectedLoaderVersion": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "ServerImportResultDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "serverId": {
            "type": "string"
          },
          "serverName": {
            "type": "string"
          },
          "imported": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          },
          "replaced": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "ErrorDTO": {
        "type": "object",
        "description": "One error envelope for every non-2xx response across every /v1/ route (P2.4 SS5-6, a D-006-point-3 correction unifying the baseline's split Error/typed-DTO failure pattern).",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable snake_case identifier, e.g. not_found, conflict, invalid_body."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, iOS-visible text."
          },
          "helpId": {
            "type": "string",
            "nullable": true,
            "description": "Optional pointer into GET /v1/help/{helpId} (P2.2)."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Optional free-form structured context."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
    }
  }
}